backend/requirements.txt
flask==2.3.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.4.4
python-dotenv==1.0.0
tensorflow==2.12.0
joblib==1.2.0
pillow==9.5.0
scikit-learn==1.2.2
mysql-connector-python==8.0.33
requests==2.31.0